Factor out some shared utilities, make tests succeed
